CODE/requirements.txt
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
Flask==1.0.2
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-vision==0.35.0
google-images-download==2.4.2
googleapis-common-protos==1.5.5
grpcio==1.16.1
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
networkx==2.2
nltk==3.3
nose==1.3.7
numpy==1.15.4
pandas==0.23.4
plac==0.9.6
preshed==2.0.1
protobuf==3.6.1
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
rake-nltk==1.0.4
regex==2018.1.10
requests==2.20.0
rsa==4.0
scipy==1.1.0
selenium==3.14.1
six==1.11.0
spacy==2.0.16
style==1.1.0
thinc==6.12.0
toolz==0.9.0
tqdm==4.28.1
ujson==1.35
update==0.0.1
urllib3==1.24
Werkzeug==0.14.1
whitenoise==4.1
wrapt==1.10.11
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz#egg=en_core_web_sm==2.0.0
